Parse genops.tab
