+++ gen/Main.hs
@@ -1,7 +1,9 @@
 import System.Process (readProcess)
 import System.Environment (getArgs)
 
-import qualified Ops
+import qualified Code
+import qualified Parse
+
 
 main :: IO ()
 main =
@@ -9,6 +11,16 @@ main =
     case commandLineArguments of
       [] ->
         error "Missing erlang version command-line argument!"
-      v : _ ->
-        do  ops <- Ops.parse <$> readProcess "curl" [ Ops.downloadUrl v ] ""
-            print ops
+      version : _ ->
+        do  opsTab <- download version "/erts/emulator/beam/ops.tab"
+            genopTab <- download version "/lib/compiler/src/genop.tab"
+            case Code.generate <$> Parse.ops opsTab <*> Parse.genop genopTab of
+              Left parseError -> error $ show parseError
+              Right haskell -> print haskell -- TODO: write to src/
+
+
+download :: String -> String -> IO String
+download version path =
+  readProcess "curl" [ rootUrl ++ version ++ path ] ""
+  where
+    rootUrl = "https://raw.githubusercontent.com/erlang/otp/"
